feat(cli): migrate proxy to new executor way
