+++ packages/cli/src/cli/commands/proxy.d.ts
@@ -1,0 +1,15 @@
+/**
+ *
+ * @param {import("@compas/stdlib").Logger} logger
+ * @param {import("../types").CliExecutorState} state
+ * @returns {import("../types").CliResult}
+ */
+export function cliExecutor(
+  logger: import("@compas/stdlib").Logger,
+  state: import("../types").CliExecutorState,
+): import("../types").CliResult;
+/**
+ * @type {import("../../generated/common/types").CliCommandDefinitionInput}
+ */
+export const cliDefinition: import("../../generated/common/types").CliCommandDefinitionInput;
+//# sourceMappingURL=proxy.d.ts.map
